+++ .coveragerc
@@ -5,6 +5,6 @@ branch = True
 omit =
   */migrations/*
   treeherder/config/wsgi.py
-  treeherder/settings/base.py
-  treeherder/settings/local.py
-  treeherder/settings/local.sample.py
+  treeherder/config/settings.py
+  treeherder/config/settings_local.py
+  treeherder/config/settings_local.example.py
