+++ manage.py
@@ -4,7 +4,7 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ["DJANGO_SETTINGS_MODULE"] = "treeherder.settings"
+    os.environ["DJANGO_SETTINGS_MODULE"] = "treeherder.config.settings"
 
     from django.core.management import execute_from_command_line
