+++ tests/conftest.py
@@ -33,7 +33,7 @@ def pytest_sessionstart(session):
 
     """
     sys.path.append(dirname(dirname(__file__)))
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.config.settings")
     from django.conf import settings
     from django.test.runner import DiscoverRunner
