+++ treeherder/celery.py
@@ -6,7 +6,7 @@ from celery import Celery
 from django.conf import settings
 
 # set the default Django settings module for the 'celery' program.
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'treeherder.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'treeherder.config.settings')
 
 app = Celery('treeherder')
