+++ treeherder/config/wsgi.py
@@ -11,8 +11,8 @@ import os
 # We defer to a DJANGO_SETTINGS_MODULE already in the environment. This breaks
 # if running multiple sites in the same mod_wsgi process. To fix this, use
 # mod_wsgi daemon mode with each site in its own daemon process, or use
-# os.environ["DJANGO_SETTINGS_MODULE"] = "webapp.settings"
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.settings")
+# os.environ["DJANGO_SETTINGS_MODULE"] = "treeherder.config.settings"
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "treeherder.config.settings")
 
 from django.core.cache.backends.memcached import BaseMemcachedCache
 from django.core.wsgi import get_wsgi_application
